Refactor SubscribeTask
